Feito o ajuste no banco e ajustado o front

+++ backend/src/main/java/com/ifpr/backend/model/ResumoMensal.java
@@ -1,11 +1,10 @@
 package com.ifpr.backend.model;
 
-import java.math.BigDecimal;
 
 public record ResumoMensal(
-    BigDecimal totalReceitas,
-    BigDecimal totalDespesas,
-    BigDecimal saldo
+    Double totalReceitas,
+    Double totalDespesas,
+    Double saldo
 ) {}
 
 // Usado record aqui porque é só um "pacote" de dados de saída, não precisa ser entidade de banco.)
